pyproject.toml
[project]
name = "ha-sync"
version = "0.1.0"
description = "Sync Home Assistant UI config (dashboards, automations, helpers) to/from local YAML files"
readme = "README.md"
authors = [
    { name = "Douwe Maan", email = "hi@douwe.me" }
]
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15",
    "aiohttp>=3.9",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "watchfiles>=1.0",
    "rich>=13.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "respx>=0.22",
]

[project.scripts]
ha-sync = "ha_sync.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ha_sync"]

[tool.pyright]
pythonVersion = "3.13"
include = ["src"]
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
